V 0.0.27 update allowing to work with country names not ready

+++ R/rangemap_bound.R
@@ -7,16 +7,31 @@
 #' in the working directory if it is needed.
 #'
 #' @param occurrences a data.frame containing species occurrences, columns must be:
-#' Species, Longitude, and Latitude.
+#' Species, Longitude, and Latitude. If not defined, adm_areas must be defined and areas specified
+#' in there will be used as part of the species range.
+#' @param adm_areas (character) a vector of names of administrative areas known to be occupied by
+#' the species, names depend on the boundary_level selected. Check \code{\link{adm_areas}} data
+#' documentation for an idea of how to define names in this parameter. If not defined, occurrences
+#' must exist.
 #' @param country_code (character) vector of country codes that will be considered when
-#' creating the species range. Note that including neighbor countries may be necessary to calculate
-#' the extent of occurrence based on the IUCN criteria, even if they are not occupied by the species.
-#' Codes follow the ISO-3166-1 norm as in function \code{\link[raster]{getData}}.
+#' creating the species range. If country selection is required to be only based on occurrences,
+#' including neighbor countries is not necessary. Use \code{\link{rangemap_explore}} for a preview
+#' of potential countries involved in the analysis. Codes follow the ISO-3166-1 norm as in function
+#' \code{\link[raster]{getData}}. If not defined, polygons must be included. Ignored if polygons
+#' are included.
 #' @param boundary_level (numeric) the level of the administrative boundaries from 0 to 5 (0 is
-#' the country and higher values equal finer divisions), default = 0.
-#' @param polygons (optional) a SpatialPolygon object that will be used instead of boundaries
-#' to create species ranges based on overlapping of species records with these layer. If defined,
-#' argument boundaries will not be considered.
+#' the country and higher values equal finer divisions), default = 0. Ignored if polygons are
+#' included.
+#' @param polygons (optional) a SpatialPolygonDataFrame object that will be used instead of boundaries
+#' especified in country_code to create species ranges based on overlapping of species records
+#' with these layer as well as names defined in adm_areas. Projection must be Geographic (longitude,
+#' latitude). If adm_areas is defined, polygons must have, as part of data, a fied (column) named
+#' "adm_areas" for selecting extra areas based on the names of defined administrative areas.
+#' If defined, arguments country_code and boundary level will be ignored.
+#' @param kept_data (logical) if TRUE and polygons is not defined, data dowloaded from the GADM data
+#' base will be kept in the working directory. Useful if all or the part of the downloaded polygons
+#' for the creation of an species range area will be used in posterior analyses because those polygons
+#' will not be downloaded again and time will be saved. Default = FALSE.
 #' @param dissolve (logical) if TRUE distint polygons selected as part of the species range will
 #' be disolved for creating simpler polygons, default = FALSE. The dissolving process may take some
 #' more time, especially if species with broad distributions are being analyzed.
@@ -28,7 +43,8 @@
 #' SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
 #' object projected to the Azimuthal equal area projection.
 #'
-#' @details Boundaries used are loaded using the \code{\link[raster]{getData}} funcion.
+#' @details Boundaries for countries defined in country_code are loaded using the
+#' \code{\link[raster]{getData}} funcion.
 #'
 #' @examples
 #' if(!require(rgbif)){
@@ -50,14 +66,18 @@
 #' occ_g <- occ[!is.na(occ$decimalLatitude) & !is.na(occ$decimalLongitude),
 #'             c("name", "decimalLongitude", "decimalLatitude")]
 #'
+#' # checking which countries may be involved in the analysis
+#' rangemap_explore(occurrences = occ_g)
+#'
 #' level <- 0
+#' adm <- "Ecuador" # Athough no record is on this country, we know it is in Ecuador
 #' dissolve <- FALSE
 #' save <- TRUE
 #' name <- "test"
 #' countries <- c("PER", "BRA", "COL", "VEN", "ECU", "GUF", "GUY", "SUR", "BOL")
 #'
-#' bound_range <- rangemap_bound(occurrences = occ_g, country_code = countries, boundary_level = level,
-#'                               dissolve = dissolve, save_shp = save, name = name)
+#' bound_range <- rangemap_bound(occurrences = occ_g, adm_areas = adm, country_code = countries,
+#'                               boundary_level = level, dissolve = dissolve, save_shp = save, name = name)
 
 # Dependencies: maps (map),
 #               maptools (map2SpatialPolygons),
@@ -67,15 +87,89 @@
 #               sp (SpatialPointsDataFrame, spTransform, SpatialPolygonsDataFrame,
 #                   CRS, over, Polygons, Polygon, SpatialPolygons, proj4string)
 
-rangemap_bound <- function(occurrences, country_code, boundary_level = 0,
-                           polygons, dissolve = FALSE, save_shp = FALSE, name) {
+rangemap_bound <- function(occurrences, adm_areas, country_code, boundary_level = 0, polygons,
+                           kept_data = FALSE, dissolve = FALSE, save_shp = FALSE, name) {
   # testing potential issues
-  if (missing(occurrences)) {
-    stop("Argument occurrences is necessary to perform the analysis")
+  if (!missing(polygons)) {
+    if (class(polygons) != "SpatialPolygonsDataFrame")
+    stop("If defined, polygons must be a SpatialPolygonsDataFrame object. If adm_areas",
+         "\nis defined, one of the fields (columns) of polygons data must be named \"adm_areas\"",
+         "\nand contain names of administrative areas for selection.", sep = "")
   }
 
-  if (dim(occurrences)[2] != 3) {
-    stop("occurrences data.frame must have the following columns: \nSpecies, Longitude, and Latitude")
+  if (missing(occurrences) & missing(adm_areas)) {
+    stop("Occurrences and/or adm_areas must exist to perform the analysis.")
+  }else {
+    if (!missing(occurrences) & missing(adm_areas)) {
+      if (dim(occurrences)[2] != 3) {
+        stop("occurrences data.frame must have the following columns: \nSpecies, Longitude, and Latitude.")
+      }
+    }
+    if (!missing(occurrences) & !missing(adm_areas)) {
+      if (dim(occurrences)[2] != 3) {
+        warning(paste("occurrences data.frame does not have required arrangment:",
+                      "\nSpecies, Longitude, and Latitude.", " Species range will be",
+                      "\ncreated using adm_areas only.", sep = ""))
+      }
+      if (!missing(polygons)) {
+        if(sum("adm_areas" %in% names(polygons@data)) == 0) {
+          stop("Data of polygons does not contain a field (column) named \"adm_names\", see help.")
+        }
+      }
+    }
+
+    if (!missing(adm_areas)) {
+      if (!missing(polygons)) {
+        if(sum("adm_areas" %in% names(polygons@data)) == 0) {
+          stop("Data of polygons does not contain a field (column) named \"adm_names\", see help.")
+        }
+
+        polynam <- polygons@data
+        a_a_names <- as.vector(unique(polynam[, "adm_areas"]))
+
+        if (sum(adm_areas %in% a_a_names) != length(adm_areas)) {
+          warning(paste("Not all of the administrative areas defined in adm_areas coincide",
+                        "\nwith the available names for administrative areas in polygons.",
+                        sep = ""))
+        }
+
+        if (sum(adm_areas %in% a_a_names) == 0 & missing(occurrences)) {
+          stop(paste("None of the administrative areas defined in adm_areas coincides",
+                     "\nwith the available names for administrative areas in polygons.",
+                     sep = ""))
+        }
+
+        if (sum(adm_areas %in% a_a_names) == 0 & !missing(occurrences)) {
+          warning(paste("None of the administrative areas defined in adm_areas coincides",
+                        "\nwith the available names for administrative areas in polygons.",
+                        "\nSpecies range will be created using occurrences only.",
+                        sep = ""))
+        }
+      }else{
+        a_names <- paste("NAME", boundary_level, sep = "_")
+        a_a_names <- as.vector(unique(adm_area_names[adm_area_names$ISO3 %in% country_code,
+                                                     a_names]))
+
+        if (sum(adm_areas %in% a_a_names) != length(adm_areas)) {
+          warning(paste("Not all of the administrative areas defined in adm_areas coincide",
+                        "\nwith the available names for that level in the countries listed",
+                        "\nin country_code.", sep = ""))
+        }
+
+        if (sum(adm_areas %in% a_a_names) == 0 & missing(occurrences)) {
+          stop(paste("None of the administrative areas defined in adm_areas coincides",
+                     "\nwith the available names for that level in the countries listed",
+                     "\nin country_code.", sep = ""))
+        }
+
+        if (sum(adm_areas %in% a_a_names) == 0 & !missing(occurrences)) {
+          warning(paste("None of the administrative areas defined in adm_areas coincides",
+                        "\nwith the available names for that level in the countries listed",
+                        "\nin country_code. Species range will be created using occurrences only.",
+                        sep = ""))
+        }
+      }
+    }
   }
 
   # erase duplicate records
@@ -86,6 +180,17 @@ rangemap_bound <- function(occurrences, country_code, boundary_level = 0,
   WGS84 <- sp::CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
   occ_sp <- sp::SpatialPointsDataFrame(coords = occ[, 2:3], data = occ,
                                        proj4string = WGS84)
+
+  # keeping only records in land
+  occ_sp <- occ_sp[!is.na(sp::over(occ_sp, as(polygons, "SpatialPolygons"))), ]
+
+  # project the points using their centriods as reference
+  centroid <- rgeos::gCentroid(occ_sp, byid = FALSE)
+
+  AEQD <- sp::CRS(paste("+proj=aeqd +lat_0=", centroid@coords[2], " +lon_0=", centroid@coords[1],
+                        " +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs", sep = ""))
+
+  occ_pr <- sp::spTransform(occ_sp, AEQD)
 
   # world map or user map fro creating species range
   if (missing(polygons)) {
@@ -99,21 +204,12 @@ rangemap_bound <- function(occurrences, country_code, boundary_level = 0,
 
     polygons@data[, 6:dim(polygons@data)[2]] <- NULL # erase non-relevant columns to avoid to much data
 
-    # erasing rds files in working directory
-    erase_rds <- list.files(path = ".", pattern = "^GADM_", full.names = TRUE)
-    unlink(erase_rds)
+    if (kept_data == FALSE) {
+      # erasing rds files in working directory
+      erase_rds <- list.files(path = ".", pattern = "^GADM_", full.names = TRUE)
+      unlink(erase_rds)
+    }
   }
-
-  # keeping only records in land
-  occ_sp <- occ_sp[!is.na(sp::over(occ_sp, as(polygons, "SpatialPolygons"))), ]
-
-  # project the points using their centriods as reference
-  centroid <- rgeos::gCentroid(occ_sp, byid = FALSE)
-
-  AEQD <- sp::CRS(paste("+proj=aeqd +lat_0=", centroid@coords[2], " +lon_0=", centroid@coords[1],
-                        " +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs", sep = ""))
-
-  occ_pr <- sp::spTransform(occ_sp, AEQD)
 
   # project polygons
   polygons <- sp::spTransform(polygons, AEQD)
@@ -125,12 +221,11 @@ rangemap_bound <- function(occurrences, country_code, boundary_level = 0,
   if (dissolve == TRUE) {
     cat("\nDissolving polygons, please wait...\n")
     polygons@data$union_field <- rep("Union", length(polygons@data[, 1])) # new field for union
-
     polygons <- rgeos::gUnaryUnion(polygons, id = polygons@data$union_field) # now dissolve
 
     boundaries@data$union_field <- rep("Union", length(boundaries@data[, 1])) # new field for union
-
     boundaries <- rgeos::gUnaryUnion(boundaries, id = boundaries@data$union_field) # now dissolve
+    boundaries <- raster::disaggregate(boundaries)
   }
 
   # calculate areas in km2
